Combined handler and color interfaces
